+++ src/core/ref/before-render.js
@@ -1,6 +1,9 @@
 /** eslint-disable */
 import { START } from '../../support/priv-constant';
 import makeObState from '../state/make-ob-state';
+import ccContext from '../../cc-context/index';
+
+const { store } = ccContext;
 
 export default function (ref) {
   const ctx = ref.ctx;
@@ -8,14 +11,20 @@ export default function (ref) {
 
   // 处于收集观察依赖
   if (ctx.__$$autoWatch) {
+    
     if (ctx.__$$hasModuleState) {
-      //每次渲染前都将最新的模块state合进来, 防止render期间读取已过期状态, 此处使用mstate，避免触发get
-      Object.assign(ref.state, ctx.mstate);
+      const { __$$prevModuleVer, module: refModule } = ctx;
+      const moduleVer = store.getModuleVer(refModule);
+      const mVer = moduleVer[refModule];
+      if (__$$prevModuleVer[refModule] !== mVer) {
+        __$$prevModuleVer[refModule] = mVer;
+        // 比较版本, 防止render期间读取已过期状态, 此处使用mstate，避免触发get
+        Object.assign(ref.state, ctx.mstate);
+      }
 
       // 每次生成的state都是一个新对象，让effect逻辑里prevState curState对比能够成立
-      ref.state = makeObState(ref, ref.state, ctx.module, true);
+      ref.state = makeObState(ref, ref.state, refModule, true);
       ctx.state = ref.state;
-      // ctx.moduleState = makeObState(ref, ctx.mstate, ctx.module, true);
 
       ctx.__$$curWaKeys = {};
       ctx.__$$compareWaKeys = ctx.__$$nextCompareWaKeys;